api_backend/requirement.txt
fastapi
uvicorn
transformers
peft
datasets
pyjwt
python-multipart
